Tweak stemming heuristics
If first result is much longer then requested key - try again with
shorter lookup key. This in some cases brings in shorter matches
which is usually better. Ideally all prefix matches would be sorted by
length, but it's probably not feasible because of key sort order in
dictionary.

Another tweak is to limit number of characters to chop off since
resulting matches are not very good if found and if long input is
provided it takes a rather long time.