Remove "internal use" blobs from the lookup results
The dictionary generators add "internal" resources to the slob file
using key names such as "~/css/default.css". These are for internal use
and should not be made visible in the lookup results.

This commit introduces a filter to remove those: any blob that has a
content type which is not HTML or plain text is filtered out. However,
since retrieving the content type requires performing I/O, for
efficiency reasons the content type check is only done if the key
starts with the prefix "~/" — any other key is assumed to be "public".
Although the standard dictionary generators do not add "public" content
under the "~/" prefix, filtering is not based only on key alone to
avoid affecting existing dictionaries that were created using a
different toolchain.